Fix recipes with missing alternatives being counted incorrectly
This could occur with oredict-based recipes.

The problem was that TE hardened glass is not the first
entry in the oredict alternatives.
When one of this was present, the second request for
this would be incorrectly counted as the first oredict
alternative would be considered unavailable in the first try,
so the second simulated extract would also (incorrectly) succeed.

Closes #33