Fix rare cases where overlapping jobs would cause halting
If a job would cause items to be added to the network,
and another job would consume those items in parallel,
the outputs of the first job could be consumed by the second job,
so that the first job sometimes would not be marked as completed.

The solution involved not allowing new crafting jobs if the observer
knows that it still needs to forcefully observe certain parts of the network.
This would be the case when another job inserts items as output into the network.

Closes #48

Small example of the problem:
Problem occurs because we already had instances in the network!
* storageidx = 10
* craft +6
* insert +3 (actual = 13)
* detect +3
* storageidx = 13
* insert +3 (actual = 16, but idx not updated yet)
* craft X and immediately consume -6 (actual = 10)
* detect -3
* storageidx = 10
==> craft +6 never finishes, because the last +3 was never detected!!!